Add intention hiding lines not containing selected substring
