+++ src/main/kotlin/com/intellij/ideolog/foldings/FoldingCalculatorTask.kt
@@ -16,6 +16,7 @@ import java.util.*
 
 val hiddenItemsKey = Key.create<HashSet<Pair<Int, String>>>("Log.HiddenColumnValues")
 val hiddenSubstringsKey = Key.create<HashSet<String>>("Log.HiddenSubStrings")
+val whitelistedSubstringsKey = Key.create<HashSet<String>>("Log.WhitelistedSubStrings")
 val whitelistedItemsKey = Key.create<HashSet<Pair<Int, String>>>("Log.WhitelistedColumnValues")
 val hideLinesAboveKey= Key.create<Int>("Log.HideLinesAbove")
 val hideLinesBelowKey= Key.create<Int>("Log.HideLinesBelow")
@@ -37,6 +38,7 @@ class FoldingCalculatorTask(project: Project, val editor: Editor, fileName: Stri
   val settings = LogHighlightingSettingsStore.getInstance()
   val hiddenItems = editor.document.getUserData(hiddenItemsKey) ?: emptySet<Pair<Int, String>>()
   val hiddenSubstrings = editor.document.getUserData(hiddenSubstringsKey) ?: emptySet<String>()
+  val whitelistedSubstrings = editor.document.getUserData(whitelistedSubstringsKey) ?: emptySet<String>()
   val whitelistedItems = editor.document.getUserData(whitelistedItemsKey) ?: emptySet<Pair<Int, String>>()
   val hideLinesAbove: Int = editor.document.getUserData(hideLinesAboveKey) ?: -1
   val hideLinesBelow: Int = editor.document.getUserData(hideLinesBelowKey) ?: Int.MAX_VALUE
@@ -155,6 +157,11 @@ class FoldingCalculatorTask(project: Project, val editor: Editor, fileName: Stri
 
     if (hiddenSubstrings.isNotEmpty()) {
       if (hiddenSubstrings.any { line.contains(it) })
+        return false
+    }
+
+    if (whitelistedSubstrings.isNotEmpty()) {
+      if (!whitelistedSubstrings.all { line.contains(it) })
         return false
     }
