+++ src/main/kotlin/com/intellij/ideolog/intentions/HideLinesNotContainingSubstringIntention.kt
@@ -1,0 +1,9 @@
+package com.intellij.ideolog.intentions
+
+import com.intellij.ideolog.foldings.whitelistedSubstringsKey
+
+class HideLinesNotContainingSubstringIntention : HideLinesIntention(whitelistedSubstringsKey) {
+
+  override fun getText() = "Hide lines not containing '$shortSelection'"
+
+}
